cache: Implement SyncDCache/InvalidDCache in C
The previous implementation looped over and read both ways of every
cache line, doing a manual comparison of the address and then doing
index invalidate on match.

This replaces that with a simple loop over each cache line using the Hit
invalidate cache ops, which will tell the hardware to check for a cache
hit and invalidate.